+++ packages/core/src/components/fields/index.ts
@@ -2,6 +2,7 @@ import { Field, FormContextType, RegistryFieldsType, RJSFSchema, StrictRJSFSchem
 
 import ArrayField from './ArrayField';
 import BooleanField from './BooleanField';
+import FallbackField from './FallbackField';
 import LayoutGridField from './LayoutGridField';
 import LayoutHeaderField from './LayoutHeaderField';
 import LayoutMultiSchemaField from './LayoutMultiSchemaField';
@@ -23,6 +24,7 @@ function fields<
     ArrayField: ArrayField as unknown as Field<T, S, F>,
     // ArrayField falls back to SchemaField if ArraySchemaField is not defined, which it isn't by default
     BooleanField,
+    FallbackField,
     LayoutGridField,
     LayoutHeaderField,
     LayoutMultiSchemaField,